+++ sdk/core/azure-core-http-test-common/src/main/java/com/azure/core/http/test/common/junitextensions/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Package containing JUnit extensions that are used by HTTP client implementations to test against.
- */
-package com.azure.core.http.test.common.junitextensions;