+++ sdk/core/azure-core-test/src/main/java/com/azure/core/test/implementation/models/PizzaSize.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.core.http.test.common.models;
+package com.azure.core.test.implementation.models;
 
 /**
  * Pizza size.