+++ sdk/core/azure-core/src/test-shared/java/com/azure/core/test/shared/LocalTestServer.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.core.http.test.common;
+package com.azure.core.test.shared;
 
 import org.eclipse.jetty.server.HttpConfiguration;
 import org.eclipse.jetty.server.HttpConnectionFactory;
@@ -189,7 +189,7 @@ public class LocalTestServer {
 
     private static byte[] fullyReadRequest(InputStream requestBody) throws IOException {
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        HttpTestUtils.copy(requestBody, outputStream);
+        CoreTestUtils.copy(requestBody, outputStream);
         return outputStream.toByteArray();
     }
 }