+++ sdk/core/azure-core/src/test-shared/java/com/azure/core/test/shared/SyncAsyncTest.java
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.core.http.test.common.junitextensions;
+
+package com.azure.core.test.shared;
 
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -13,7 +14,10 @@ import java.util.concurrent.Callable;
 
 /**
  * Indicates that test should be branched out into sync and async branch resulting in two separate test cases.
- * The {@link SyncAsyncExtension#execute(Callable, Callable)} should be used in the test to branch out.
+ * The {@link SyncAsyncExtension#execute(Callable, Callable)} should be used in the test
+ * to branch out.
+ * Using azure-core copy of the com.azure.core.test.annotation.SyncAsyncTest class
+ * since azure-core cannot take dependency on azure-core-test package.
  */
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)